feat(transformations): Add `Tranformation` class
The `Tranformation` struct and its methods are added and tested.
The fuctions `rotation_*`, `traslation` and `scale` are intended
to be used as creators for elementary transformations.
Attention: note that a `0.0` parameter given to the `scale` function
generates a panic due to division by zero.

The `Matrix` struct's elements are switched from being
a plain array to an array of array,
also they are now correctly representing a 4x4 matrix
instead of 3x3