reformat!(render): move `&mut` away from renderers
the trait `Solve` now not require `&mut self` that was needed
cause some renderers (pathtracer) had a field `Pcg` that must be mutable
to work correctly. Now is the trait `Solve` that contain a `&mut Pcg` so
the logic is shifted to the trait that will solve the rendering
equation, and no more to the renderers.
This reformat is also necessary to permit to easily parallelize the
`fire_all_rays` routine in `imagetracer.rs`